requirements.txt
azureml==0.2.7
azureml_core==1.52.0
einops==0.7.0
fvcore==0.1.5.post20221221
hydra-core==1.3.2
matplotlib==3.7.1
numpy==1.25.0
omegaconf==2.3.0
opencv_python==4.8.0.74
opencv_python_headless==4.8.0.74
Pillow==10.1.0
pycocotools==2.0.6
pytorch_lightning==2.0.4
scikit_learn==1.3.0
scipy==1.11.3
submitit==1.5.0
timm==0.9.2
tqdm==4.65.0
wget==3.2
tensorboard==2.13.0
git+https://github.com/lucasb-eyer/pydensecrf.git
